+++ optarena/benchmarks/ml/lenet/lenet_numpy.py
@@ -18,8 +18,7 @@ def conv2d(input, weights):
     for i in range(H_out):
         for j in range(W_out):
             output[:, i, j, :] = np.sum(
-                input[:, i:i + K, j:j + K, :, np.newaxis] *
-                weights[np.newaxis, :, :, :],
+                input[:, i:i + K, j:j + K, :, np.newaxis] * weights[np.newaxis, :, :, :],
                 axis=(1, 2, 3),
             )
 
@@ -28,20 +27,15 @@ def conv2d(input, weights):
 
 # 2x2 maxpool operator, as used in LeNet-5
 def maxpool2d(x):
-    output = np.empty(
-        [x.shape[0], x.shape[1] // 2, x.shape[2] // 2, x.shape[3]],
-        dtype=x.dtype)
+    output = np.empty([x.shape[0], x.shape[1] // 2, x.shape[2] // 2, x.shape[3]], dtype=x.dtype)
     for i in range(x.shape[1] // 2):
         for j in range(x.shape[2] // 2):
-            output[:, i, j, :] = np.max(x[:, 2 * i:2 * i + 2,
-                                          2 * j:2 * j + 2, :],
-                                        axis=(1, 2))
+            output[:, i, j, :] = np.max(x[:, 2 * i:2 * i + 2, 2 * j:2 * j + 2, :], axis=(1, 2))
     return output
 
 
 # LeNet-5 Convolutional Neural Network (inference mode)
-def lenet5(input, conv1, conv1bias, conv2, conv2bias, fc1w, fc1b, fc2w, fc2b,
-           fc3w, fc3b, N, C_before_fc1, out):
+def lenet5(input, conv1, conv1bias, conv2, conv2bias, fc1w, fc1b, fc2w, fc2b, fc3w, fc3b, N, C_before_fc1, out):
     x = relu(conv2d(input, conv1) + conv1bias)
     x = maxpool2d(x)
     x = relu(conv2d(x, conv2) + conv2bias)
